# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seed Server",
    "version": "0.5"
  },
  "paths": {
...
    "/v1/actions": {
      "post": {
        "summary": "Actions",
        "operationId": "actions_v1_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs/submit": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit Job",
        "description": "Submit a background job for async processing.\n\n**Supported Job Types:**\n- `lesson_generate`: Generate a lesson (params = LessonGenerateRequest)\n- `diagnostic_generate`: Generate diagnostic items (params = DiagnosticGenerateRequest)\n\n**Queues:**\n- `q_fast`: High-priority queue (< 5s typical wait)\n- `q_batch`: Standard queue (< 30s typical wait)\n- `q_low`: Low-priority queue (< 2min typical wait)\n\n**Example:**\n```json\n{\n  \"job_type\": \"lesson_generate\",\n  \"params\": {\n    \"mode\": \"vocabulary\",\n    \"target_lang\": \"Spanish\",\n    \"native_lang\": \"English\",\n    \"level\": \"A2\",\n    \"lesson_length\": 5\n  },\n  \"priority\": 10,\n  \"queue\": \"q_fast\"\n}\n```\n\n**Returns:**\n- `job_id`: Use this to poll status or stream updates\n- `status`: Initial status (always \"queued\")\n- `estimated_wait_sec`: Rough estimate based on queue depth",
        "operationId": "submit_job_v1_jobs_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/status/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Status",
        "description": "Get current status of a job.\n\n**Polling Pattern:**\n```javascript\nasync function waitForJob(jobId) {\n    while (true) {\n        const status = await fetch(`/v1/jobs/status/${jobId}`).then(r => r.json());\n        \n        if (status.status === 'done') {\n            return status.result;\n        }\n        if (status.status === 'failed') {\n            throw new Error(status.error);\n        }\n        \n        // Poll every 2 seconds\n        await new Promise(r => setTimeout(r, 2000));\n    }\n}\n```\n\n**Better: Use streaming endpoint** `/jobs/status/{job_id}/stream` for real-time updates without polling.",
        "operationId": "get_job_status_v1_jobs_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/status/{job_id}/stream": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Stream Job Status",
        "description": "Stream job status updates in real-time using SSE.\n\n**Events:**\n- `status`: Status update (queued -> running -> done/failed)\n- `progress`: Progress information (if available)\n- `complete`: Job finished with result\n- `error`: Job failed\n\n**Client Example:**\n```javascript\nconst eventSource = new EventSource(`/v1/jobs/status/${jobId}/stream`);\n\neventSource.addEventListener('status', (e) => {\n    const data = JSON.parse(e.data);\n    console.log('Status:', data.status);\n});\n\neventSource.addEventListener('complete', (e) => {\n    const data = JSON.parse(e.data);\n    console.log('Result:', data.result);\n    eventSource.close();\n});\n```\n\nThis is more efficient than polling and provides instant updates.",
        "operationId": "stream_job_status_v1_jobs_status__job_id__stream_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/list": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List user's jobs with optional filtering.\n\n**Parameters:**\n- `limit`: Max results (default 20, max 100)\n- `status`: Filter by status (queued, running, done, failed)\n\n**Returns:**\nArray of job summaries with status and timestamps.",
        "operationId": "list_jobs_v1_jobs_list_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Action": {
        "type": "string",
        "enum": [
          "fix",
          "translate",
          "summarize",
          "ask"
        ],
        "title": "Action"
      },
      "ActionRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "personaId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaid"
          }
        },
        "type": "object",
        "required": [
          "action",
          "text"
        ],
        "title": "ActionRequest"
      },
      "ActionResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "eta_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Hint"
          },
          "result_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Text"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "personaIdUsed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaidused"
          },
          "fallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackreason"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "mode"
        ],
        "title": "ActionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "queue_name": {
            "type": "string",
            "title": "Queue Name"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "not_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Not Before"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "personaIdUsed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaidused"
          },
          "fallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackreason"
          },
          "result_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Text"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "action",
          "mode",
          "status",
          "queue_name",
          "priority",
          "provider",
          "model",
          "created_at"
        ],
        "title": "JobResponse"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "done",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobStatusResponse",
        "description": "Job status information"
      },
      "JobSubmitRequest": {
        "properties": {
          "job_type": {
            "type": "string",
            "title": "Job Type",
            "description": "Job type: lesson_generate, diagnostic_generate"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Job parameters (varies by type)"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "description": "Job priority (higher = faster)",
            "default": 0
          },
          "queue": {
            "type": "string",
            "title": "Queue",
            "description": "Target queue: q_fast, q_batch, q_low",
            "default": "q_fast"
          }
        },
        "type": "object",
        "required": [
          "job_type",
          "params"
        ],
        "title": "JobSubmitRequest",
        "description": "Request to submit a background job"
      },
      "JobSubmitResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "queue": {
            "type": "string",
            "title": "Queue"
          },
          "estimated_wait_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Sec"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "queue"
        ],
        "title": "JobSubmitResponse",
        "description": "Response after submitting job"
      },
...
      "Mode": {
        "type": "string",
        "enum": [
          "fast",
          "hybrid",
          "batch"
        ],
        "title": "Mode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}